feat(salesforce): add dropdown to update / create actions
