+++ packages/pieces/salesforce/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-salesforce",
-  "version": "0.0.7"
+  "version": "0.0.8"
 }
