+++ packages/pieces/salesforce/src/index.ts
@@ -5,6 +5,8 @@ import { newRecord } from './lib/trigger/new-record';
 import { newOrUpdatedRecord } from './lib/trigger/new-updated-record';
 import { upsertByExternalId } from './lib/action/upsert-by-external-id';
 import { runQuery } from './lib/action/run-sf-query';
+import { createNewObject } from './lib/action/create-new-object';
+import { UpdateObjectById } from './lib/action/update-object-by-id';
 
 export const salesforce = createPiece({
   name: 'salesforce',
@@ -12,10 +14,13 @@ export const salesforce = createPiece({
   logoUrl: 'https://cdn.activepieces.com/pieces/salesforce.png',
   version: packageJson.version,
   authors: [
-    "abuaboud"
+    "abuaboud",
+    "HKudria"
   ],
   actions: [
     runQuery,
+    createNewObject,
+    UpdateObjectById,
     upsertByExternalId
   ],
   triggers: [newRecord, newOrUpdatedRecord],
